pyproject.toml
[project]
name = "pyjolt"
version = "0.9"
description = "A batteries included async-first python webframework"
readme = "README.md"
authors = [
    { name = "MarkoSterk", email = "marko_sterk@hotmail.com" }
]
requires-python = ">=3.12"

dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "anyio>=4.8.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "certifi>=2024.12.14",
    "cffi>=1.17.1",
    "click>=8.1.8",
    "cryptography>=44.0.0",
    "greenlet>=3.1.1",
    "h11>=0.14.0",
    "httpcore>=1.0.7",
    "httpx>=0.28.1",
    "idna>=3.10",
    "jinja2>=3.1.5",
    "mako>=1.3.8",
    "markupsafe>=3.0.2",
    "marshmallow-jsonschema>=0.13.0",
    "marshmallow>=3.25.1",
    "packaging>=24.2",
    "pycparser>=2.22",
    "python-dotenv>=1.0.1",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
    "werkzeug>=3.1.3",
    "setuptools>=75.8.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "asgi-lifespan>=2.1.0",
    "pyjwt>=2.10.1",
    "redis>=4.2,<5.0",
    "websockets>=14.2",
    "motor>=3.7.0",
    "aiohttp>=3.11.12",
    "pydantic>=2.11.3",
    "requests>=2.32.3",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
scheduler = ["apscheduler>=3.11.0"]
desktop_ui = ["pywebview>=5.4", "pystray>=0.19.5", "pillow>=11.1.0"]
cache = ["redis>=4.2,<5.0"]
ai_interface = ["openai>=1.61.1", "docstring-parser>=0.16", "numpy>=2.2.2", "torch>=2.6.0", "sentence-transformers>=3.4.1", "pgvector>=0.3.6"]
rate_limiter = ["redis>=4.2,<5.0"]
full = ["apscheduler>=3.11.0", "pywebview>=5.4", "pystray>=0.19.5", "pillow>=11.1.0", "redis>=4.2,<5.0", "openai>=1.61.1", "docstring-parser>=0.16", "numpy>=2.2.2", "torch>=2.6.0", "sentence-transformers>=3.4.1", "pgvector>=0.3.6"]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
]

[project.scripts]
pyjolt = "pyjolt:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/MarkoSterk/PyJolt"
Issues = "https://github.com/MarkoSterk/PyJolt/issues"
